refactor: 프론트 로컬 QR 생성용 서비스 리팩토링
