+++ src/main/java/com/stamp/api/attendance/controller/AttendanceControllerLocal.java
@@ -1,7 +1,7 @@
 package com.stamp.api.attendance.controller;
 
 import com.stamp.api.attendance.dto.response.QRCodeRes;
-import com.stamp.api.attendance.service.LocalAttendanceService;
+import com.stamp.api.attendance.service.frontlocal.LocalAttendanceService;
 import com.stamp.global.response.ApplicationResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
